Create GraphML with Community Detection Output
Use omega score to compare the results of algorithms against each
other. Produces a csv file that has a matrix of omega scores of the
results of one community detection algorithm against another. In cases
where ground truth is present the diagonal represents the omega score
comparing against ground truth. When there is no ground truth data it
is set to 0