Bug Fix + Adding Omega vs. Runtime
Reflected change in omega score location in the JSON file in this script

Added ability to limit metrics to only a specific dataset.
Added a plot of runtime vs. accuracy. Still needs some work to make it
nicer (algorithm type should be represented by color and not by text
label).